make: Add a Makefile target that runs golangci-lint
Adds a new make target 'make verify' that is meant to run all the
linters and code checks we'll have. For now, there's a single dependency
`make verify-go-lint` that runs the golangci-lint linter.

The linter is set to use go 1.17 on purpose, running it with 1.18
wouldn't produce many results as only a handful of linters have been
ported to 1.18. It's suggested to use something like "gvm" to manage
multiple go installations.

golangci-lint is installed into the tools/ subdirectory where many of
our other tools live.

Several linters are disabled from the get go, their results are too
irritating.