Add a makefile target for `deploy`
Add the ability to deploy the compliance service using kustomize and an
existing Kubernetes cluster.

Additionally patches will add more functionality to kustomize to support
deploying the actual compliance service. This patch will be referenced
in the openshift/release configuration to ensure kustomize is used to
deploy the application in CI.

Fixes #43