Use db instead of conn for database variable
The sql module has a type for databases (DB) and a type for connections
(Conn). We referred to the database, which is returned by sql.Open(), as
a connection, but it's really of type DB.

For test writers this can be confusing if they think they're dealing
with a Conn type instead of a DB.

Let's update the variable name to more accurately reflect what
sql.Open() returns.