Implement a foreign key for subjects
We want to be able to nest subjects and build a hierarchical
relationship between them. This commit does that by adding a database
migration that adds a new column for parent_id and adds a FK
relationship to it.

This approach was discussed in https://github.com/rhmdnd/compserv/issues/84