Add a table for profiles
This commit adds a table for profiles so that we can reference controls,
and results to a parent profile.

This table is not designed, at least initially, to store more than just
a profile name and some metadata. This area will be tricky to navigate
moving forward since it would be idea to rely on OSCAL defintions for
profiles, and use OSCAL profiles distributed by standards bodies, and
not maintain them in a service.

This approach is being discussed in issue #84.

Fixes #32