Add a table for controls
This table adds a basic schema for controls, which will be referenced by
results.

Ideally, this control information will supplement information from other
sources like OSCAL catalogs, which should be the preferred source of
information, since it comes from the profile/standard authors.

This schema is being discussed in issue #84.