Document the database schema
Add a make target for generating the schema from the migrations and
including it in the database documentation.

This is useful for having a holistic view of the database, instead of
sifting through migrations to understand the relationships between all
tables. The included .sql file is only for documentation purposes.

Fixes #84