Add support for deploying app in kubernetes
This commit adds kustomize support for deploying compserv and a postgres
database on Kubernetes. Additionally, compserv can be configured to use
AWS RDS for the database.

Future changes will break apart the kustomize files into separate
overlaps, so that we have a dedicated overlay for development and leave
space for other environments. Mainly focusing on organizing the
deployment scripts so they're easy to follow.