Simplify configuration handling
We use viper to parse configuration. Previously, we would use the
library to set defaults, parse configuration, and lookup values. Then,
we'd put that into a map for other parts of the compliance service to
use.

This just introduced a middle layer in configuration handling that
wasn't necessary. Instead, let's just pass back an instance of the viper
configuration handler and let parts of the application call the values
they need out of it using `viper.GetString()`.

This is less error prone since it doesn't require maintainers to keep a
extra mental mapping between viper configuration and an internal map.