Explicitly ignore builds and tools
Git ignore is greedy, meaning `tools/` will match any subdirectory in
the repository, not just the `tools/` we use at the root. This is a
problem if we have a lot of dependencies that might have a directory in
their package called `tools/`, or anything else we ignore.

Let's be explicit in ignoring our own tools and builds so that we don't
accidentally ignore something a dependency needs.