Add support for retrieving secrets from kubernetes
We want the ability to deploy compser on a kubernetes cluster, which
also means secrets (e.g., database passwords) would be held as
kubernetes secrets.

Let's add a method to config handling that uses kubernetes clients to
fetch secrets. We'll use this in subsequent patches that use kustomize
to deploy compserv and a database on kubernetes.