Disable dependabot
We haven't been working on the project recently (within the last 6 - 9
months).

While dependabot is certainly useful, we stand up OpenShift clusters for
CI every time it proposes a dependency update.

Let's disable dependabot for the time being and we can re-enable it when
we start spending more cycles on it.